requirements.txt
albumentations==0.3.1
numpy==1.17.0
tqdm==4.40.0
matplotlib==3.1.1
scikit_image==0.15.0
openslide_python==1.1.1
opencv_python_headless==4.1.0.25
pandas==0.25.1
config==0.5.0.post0
imutils==0.5.3
optimizers==v1.9
Pillow==8.0.1
scikit_learn==0.23.2
skimage==0.0
spams==2.6.2.5
tensorboardX==2.1
torch==1.7.1
torchvision==0.8.2
